Moves testinfra staging tests into molecule/
Using a shared test directory, outside of a single scenario, so that the
same staging config tests can be referenced by multiple scenarios (we
already have virtualbox, libvirt, and AWS staging scenarios).

Updates some of the hardcoded relative paths in the OSSEC tests, as
well. Other than that small change, this is a simple move operation.