Concatenates testinfra staging vars
We're eliminating use of the TESTINFRA_TARGET_HOST env var for importing
vars. The vast majority of staging vars were intended as overrides
vis-a-vis prod; not between app and mon. We're not currently testing
prod, so that's an unnecessary optimization.

The staging.yml file is concatenated from the {app,mon}-staging.yml
files, with a few exceptions:

  * removed unused iptables ruleset (Jinja2 templates are used)
  * removed mention of SSH tor service (untested in mon-staging)