Renames testinfra files for uniqueness
Pytest expects the collected test files to be uniquely named, in terms
of basename. We didn't encounter this before because we were running the
config tests sequentially, against each host, so the collection was not
aggregated. When trying to run against a combined hostgroup, the
following error is shown:

    ==================================== ERRORS ====================================
    ____________ ERROR collecting testinfra/staging/mon/test_network.py ____________
    import file mismatch:
    imported module 'test_network' has this __file__ attribute:
      /home/conor/freedomofpress/securedrop/molecule/testinfra/staging/app/test_network.py
    which is not the same as the test file we want to collect:
      /home/conor/freedomofpress/securedrop/molecule/testinfra/staging/mon/test_network.py
    HINT: remove __pycache__ / .pyc files and/or use a unique basename for your test file modules
    =============================== warnings summary ===============================

Renaming the test files for unique basenames resolves, and is also more explicit.